+++ Bookkeeping/src/graphics/elements/ellipse.h
@@ -1,0 +1,9 @@
+#pragma once
+#include "src/core/color.h"
+
+struct Ellipse {
+    float cx, cy, rx, ry;
+    Color fill;
+    Color stroke;
+    float stroke_width;
+};