+++ Bookkeeping/src/graphics/elements/line.h
@@ -1,0 +1,8 @@
+#pragma once
+#include "src/core/color.h"
+
+struct Line {
+    float x1, y1, x2, y2;
+    Color color;
+    float width;
+};